Update aPracticeInDeath2.0
I revised the gameplay. The objective is still to get to earth but the distance one travels is not up to the player, but the game. In this version, the distance one travels to earth every time the player chooses to travel is 20 lightyears.

This version removed the options to defend against the aliens with a shield and requires the player to choose between fighting the aliens or fleeing from them. If the player flee the aliens, they go back 10 lightyears. If they fight, the computer generates a random number between 1-20, and if it is greater-than or equal-to 13, the player lives and can continue their journey.

In this version, what happens if the player chooses not to travel at any point in the game has not been programmed in yet.